fix(inmem): remove view service
